Cleanup: Get rid of duplicate getChannelName()
Get rid of duplicate getChannelName function. Use
the one with uint8_t channelNum.

Tested-by:
1. Verified RMCP+ session establishement works with newly
created user

+++ user_channel/channel_layer.cpp
@@ -137,7 +137,7 @@ ipmi_ret_t getChannelEnabledAuthType(const uint8_t chNum, const uint8_t priv,
                                                               authType);
 }
 
-std::string getChannelName(const int chNum)
+std::string getChannelName(const uint8_t chNum)
 {
     return getChannelConfigObject().getChannelName(chNum);
 }
